Improved type conversion for the option file reading

+++ ambilight/src/Options.h
@@ -19,14 +19,14 @@ public:
 
 	std::string getPortName() const;
 	std::vector<Coordinates> getCoordinates() const;
-	float getSmoothing() const;
+	double getSmoothing() const;
 	ColorGrader getColorGrader() const;
 
 private:
 	static nlohmann::json getJson(const std::string & path);
 	static std::string getPortName(const nlohmann::json & json);
 	static std::vector<Coordinates> getCoordinates(const nlohmann::json & json);
-	static float getSmoothing(const nlohmann::json & json);
+	static double getSmoothing(const nlohmann::json & json);
 	static ColorGrader getColorGrader(const nlohmann::json & json);
 	static RGBLUT getRGBLut(const nlohmann::json & json);
 	static LUT getLut(const nlohmann::json & json);
@@ -35,7 +35,7 @@ private:
 	const nlohmann::json json;
 	const std::string portName;
 	const std::vector<Coordinates> coordinates;
-	const float smoothing;
+	const double smoothing;
 	const ColorGrader colorGrader;
 
 };
